+++ CHANGELOG.md
@@ -3,6 +3,13 @@ TODO
 - support nullable types with strictNullchecks
 - type "becomes" to the property value
 
+v0.8.1
+
+- reverted callback functions in makeRules to be part of one object, so dynamic rules use the same options object (because dynamic rules return an object, so it can't return two parameters)
+- before: effect(callback, {check:{prop: "propName"}}
+- after: effect({run: callback, check:{prop: "propName"}}
+- added support for new syntax for dynamic rules
+
 v0.8.0
 
 - added a simpler syntax for adding rules,